feat(reader): show two facing pages on a wide landscape window
A page area at least 840dp wide and wider than tall now draws two slots, each the width the render window was priced at, with its page number in the outer corner and the position read as a range. The pager moves by spread, a Two pages item in the menu turns it off while the window qualifies, and a lone last page keeps its own slot.

A pinch or double tap names the page under the fingers, so zooming opens that page alone and returning to the fitted scale brings the spread back. Selection, search highlights and recognition status work in both slots, and a selection in one replaces the other. Outside a spread nothing changed.

+++ app/src/androidTest/java/com/folium/reader/reader/HorizontalReaderScreenTest.kt
@@ -174,7 +174,7 @@ class HorizontalReaderScreenTest {
         onSearchNext: () -> Unit = {},
         onSearchOcrPause: () -> Unit = {},
         onSearchOcrResume: () -> Unit = {},
-        onOcrRetry: () -> Unit = {}
+        onOcrRetry: (Int) -> Unit = {}
     ) {
         shown.value = state
         compose.setContent {
